requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.34.137
botocore==1.34.137
celery==5.4.0
Django==4.2.13
django-cron==0.6.0
django-storages==1.14.3
djangorestframework==3.15.2
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==26.0.0
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis>=3.5,<4.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
